+++ src/main/java/com/andersmmg/cityessentials/sounds/ModSounds.java
@@ -15,6 +15,7 @@ public class ModSounds {
     public static final SoundEvent MAILBOX_CLOSE = registerSoundEvent("mailbox_close");
     public static final SoundEvent ENVELOPE_OPEN = registerSoundEvent("envelope_open");
     public static final SoundEvent BELL_DING = registerSoundEvent("bell_ding");
+    public static final SoundEvent WALLET_OPEN = registerSoundEvent("wallet_open");
 
     private static SoundEvent registerSoundEvent(String name) {
         Identifier id = CityEssentials.id(name);